Move routes to own file
